[feat] Extract interpreter call helpers
