+++ src/interpreter/interpreter.c
@@ -23,7 +23,7 @@
 #include "utils/utils.h"
 #include "types/type_registry.h"
 
-static CallStack call_stack;
+CallStack call_stack;
 static bool break_flag = false;
 static bool continue_flag = false;
 
@@ -34,183 +34,6 @@ typedef enum
     ANNOTATION_TARGET_CLASS,
     ANNOTATION_TARGET_ASSIGNMENT
 } AnnotationTargetType;
-
-static Value run_async_task(AsyncTask *task)
-{
-    Function *fn = task->fn;
-    Env *env = env_create(fn->env);
-    int offset = 0;
-    if (task->has_self)
-    {
-        set_variable(env, fn->params[0], task->self);
-        offset = 1;
-    }
-    for (int i = 0; i < task->arg_count; ++i)
-        set_variable(env, fn->params[i + offset], task->args[i]);
-
-    CallFrame frame = {.env = env, .return_ptr = NULL, .returning = false};
-    push_frame(&call_stack, frame);
-    Value result = run_ast(fn->body, fn->body_count);
-    pop_frame(&call_stack);
-    env_release(env);
-    return result;
-}
-
-static Value create_async_promise(Function *fn, Value *args, int arg_count, bool has_self, Value self, int line, int column)
-{
-    AsyncTask *task = async_task_create(fn, args, arg_count, has_self, self, line, column);
-    Promise *promise = promise_create_with_task(task);
-    Value promise_val = {.type = VAL_PROMISE, .promise = promise};
-    return promise_val;
-}
-
-static Value await_promise(Value awaited, int line, int column)
-{
-    Value current = clone_value(&awaited);
-    while (current.type == VAL_PROMISE)
-    {
-        Promise *promise = current.promise;
-        PromiseState state = promise_state(promise);
-
-        if (state == PROMISE_PENDING)
-        {
-            AsyncTask *task = promise_take_task(promise);
-            if (task)
-            {
-                Value result = run_async_task(task);
-                promise_resolve(promise, result);
-                free_value(result);
-                async_task_free(task);
-                state = promise_state(promise);
-            }
-            else
-            {
-                free_value(current);
-                log_script_error(line, column, "Promise is still pending");
-                exit(1);
-            }
-        }
-
-        if (state == PROMISE_FULFILLED)
-        {
-            Value next = promise_clone_result(promise);
-            if (next.type == VAL_PROMISE && next.promise == promise)
-            {
-                free_value(next);
-                free_value(current);
-                log_script_error(line, column, "Promise resolved with itself");
-                exit(1);
-            }
-            free_value(current);
-            current = next;
-            continue;
-        }
-
-        if (state == PROMISE_REJECTED)
-        {
-            Value reason = promise_clone_reason(promise);
-            free_value(current);
-            if (reason.type == VAL_STRING)
-                log_script_error(line, column, "Promise rejected: %s", reason.str);
-            else
-                log_script_error(line, column, "Promise rejected");
-            free_value(reason);
-            exit(1);
-        }
-
-        break;
-    }
-
-    return current;
-}
-
-static Value call_value(Value callee, Value *args, int arg_count, int line, int column)
-{
-    if (callee.type == VAL_BOUND_METHOD)
-    {
-        Function *fn = callee.bound->func;
-        if (fn->param_count - 1 != arg_count)
-        {
-            log_script_error(line, column,
-                             "Function expects %d arguments, but got %d",
-                             fn->param_count - 1, arg_count);
-            exit(1);
-        }
-        Value self_val = {.type = VAL_INSTANCE, .instance = callee.bound->self};
-        if (fn->is_async)
-            return create_async_promise(fn, args, arg_count, true, self_val, line, column);
-        Env *env = env_create(fn->env);
-        set_variable(env, fn->params[0], self_val);
-        for (int p = 0; p < arg_count; ++p)
-            set_variable(env, fn->params[p + 1], args[p]);
-        CallFrame frame = {.env = env, .return_ptr = NULL, .returning = false};
-        push_frame(&call_stack, frame);
-        Value result = run_ast(fn->body, fn->body_count);
-        pop_frame(&call_stack);
-        Value ret_val = clone_value(&result);
-        env_release(env);
-        return ret_val;
-    }
-    if (callee.type == VAL_TYPE && promise_type_is_namespace(callee.cls))
-    {
-        log_script_error(line, column, "Promise cannot be instantiated directly");
-        exit(1);
-    }
-    if (callee.type == VAL_TYPE)
-    {
-        Instance *inst = instance_create(callee.cls);
-        Value inst_val = {.type = VAL_INSTANCE, .instance = inst};
-        Value init = value_get_attr(inst_val, "init");
-        if (init.type == VAL_BOUND_METHOD)
-        {
-            Function *fn = init.bound->func;
-            if (fn->param_count - 1 != arg_count)
-            {
-                log_script_error(line, column, "init expects %d arguments, got %d",
-                                 fn->param_count - 1, arg_count);
-                exit(1);
-            }
-            Env *env = env_create(fn->env);
-            Value selfv = {.type = VAL_INSTANCE, .instance = init.bound->self};
-            set_variable(env, fn->params[0], selfv);
-            for (int p = 0; p < arg_count; ++p)
-                set_variable(env, fn->params[p + 1], args[p]);
-            CallFrame frame = {.env = env, .return_ptr = NULL, .returning = false};
-            push_frame(&call_stack, frame);
-            run_ast(fn->body, fn->body_count);
-            pop_frame(&call_stack);
-            env_release(env);
-        }
-        return inst_val;
-    }
-    if (callee.type != VAL_FUNCTION)
-    {
-        log_script_error(line, column, "Attempting to call non-function");
-        exit(1);
-    }
-    Function *fn = callee.func;
-    if (fn->param_count != arg_count)
-    {
-        log_script_error(line, column, "Function expects %d arguments, but got %d",
-                         fn->param_count, arg_count);
-        exit(1);
-    }
-    if (fn->is_async)
-    {
-        Value undef = {.type = VAL_UNDEFINED};
-        return create_async_promise(fn, args, arg_count, false, undef, line, column);
-    }
-    Env *env = env_create(fn->env);
-    for (int p = 0; p < arg_count; ++p)
-        set_variable(env, fn->params[p], args[p]);
-    CallFrame frame = {.env = env, .return_ptr = NULL, .returning = false};
-    push_frame(&call_stack, frame);
-    Value result = run_ast(fn->body, fn->body_count);
-    pop_frame(&call_stack);
-    Value ret_val = clone_value(&result);
-    env_release(env);
-    return ret_val;
-}
 
 static Value exec_func_call(ASTNode *n);
 static Value resolve_attr_prefix(ASTNode *attr_node, int count);
@@ -358,7 +181,7 @@ static Value eval_node(ASTNode *n)
     case NODE_AWAIT:
     {
         Value awaited = eval_node(n->children[0]);
-        Value resolved = await_promise(awaited, n->line, n->column);
+        Value resolved = interpreter_await(awaited, n->line, n->column);
         return resolved;
     }
     case NODE_OBJECT_LITERAL:
@@ -1112,7 +935,7 @@ static Value exec_func_call(ASTNode *n)
                 for (int p = 0; p < n->child_count; ++p)
                     arg_vals[p] = eval_node(n->children[p]);
             }
-            Value promise_val = create_async_promise(fn, arg_vals, n->child_count, true, self_val, n->line, n->column);
+            Value promise_val = interpreter_create_async_promise(fn, arg_vals, n->child_count, true, self_val, n->line, n->column);
             if (arg_vals)
             {
                 for (int p = 0; p < n->child_count; ++p)
@@ -1197,7 +1020,7 @@ static Value exec_func_call(ASTNode *n)
             for (int p = 0; p < n->child_count; ++p)
                 arg_vals[p] = eval_node(n->children[p]);
         }
-        Value promise_val = create_async_promise(fn, arg_vals, n->child_count, false, undef, n->line, n->column);
+        Value promise_val = interpreter_create_async_promise(fn, arg_vals, n->child_count, false, undef, n->line, n->column);
         if (arg_vals)
         {
             for (int p = 0; p < n->child_count; ++p)
@@ -1297,7 +1120,7 @@ static bool apply_annotations(ASTNode *node, Value *value, AnnotationTargetType 
                 for (int j = 0; j < ann->arg_count; ++j)
                     args[j] = eval_node(ann->args[j]);
             }
-            Value intermediate = call_value(decorator_callable, args, ann->arg_count, ann->line, ann->column);
+            Value intermediate = interpreter_call_value(decorator_callable, args, ann->arg_count, ann->line, ann->column);
             free_value(decorator_callable);
             decorator_callable = intermediate;
             if (args)
@@ -1311,7 +1134,7 @@ static bool apply_annotations(ASTNode *node, Value *value, AnnotationTargetType 
 
         Value decorator_args[1];
         decorator_args[0] = *value;
-        Value decorated = call_value(decorator_callable, decorator_args, 1, ann->line, ann->column);
+        Value decorated = interpreter_call_value(decorator_callable, decorator_args, 1, ann->line, ann->column);
         free_value(decorator_callable);
         *value = decorated;
     }
@@ -1346,7 +1169,7 @@ static bool apply_annotations(ASTNode *node, Value *value, AnnotationTargetType 
         Value args[2];
         args[0] = *value;
         args[1] = info_val;
-        Value result = call_value(modifiers[i], args, 2, ann->line, ann->column);
+        Value result = interpreter_call_value(modifiers[i], args, 2, ann->line, ann->column);
         free_value(modifiers[i]);
         modifiers[i].type = VAL_UNDEFINED;
 
@@ -1572,7 +1395,7 @@ Value run_ast(ASTNode **nodes, int count)
                     log_script_error(n->line, n->column, "Object is not iterable");
                     exit(1);
                 }
-                Value iterator = call_value(iter_func, NULL, 0, n->line, n->column);
+                Value iterator = interpreter_call_value(iter_func, NULL, 0, n->line, n->column);
                 while (1)
                 {
                     Value next_f = value_get_attr(iterator, "__next__");
@@ -1582,7 +1405,7 @@ Value run_ast(ASTNode **nodes, int count)
                                          "Iterator missing __next__ method");
                         exit(1);
                     }
-                    Value item = call_value(next_f, NULL, 0, n->line, n->column);
+                    Value item = interpreter_call_value(next_f, NULL, 0, n->line, n->column);
                     if (item.type == VAL_UNDEFINED)
                         break;
                     set_variable(interpreter_current_env(), n->data.loop.loop_var,
